Cleaned up code

+++ safecollections/stack_bench_test.go
@@ -3,11 +3,11 @@ package safecollections
 import "testing"
 
 func BenchmarkStack_Push(b *testing.B) {
-	fillStack(b.N)
+	fillStack(uint64(b.N))
 }
 
 func BenchmarkStack_Pop(b *testing.B) {
-	stack := fillStack(b.N)
+	stack := fillStack(uint64(b.N))
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		stack.Pop()
